Initializing cucumber adn first test #4
