+++ pom.xml
@@ -15,9 +15,15 @@ http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <dependencies>
         <!-- Other dependencies -->
         <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter-engine</artifactId>
-            <version>5.9.3</version>
+            <groupId>io.cucumber</groupId>
+            <artifactId>cucumber-java</artifactId>
+            <version>7.12.0</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.cucumber</groupId>
+            <artifactId>cucumber-junit</artifactId>
+            <version>7.12.0</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
