+++ src/test/resources/cucumber.feature
@@ -1,0 +1,7 @@
+Feature: Is it Szakest yet?
+  Everybody wants to know when it's over
+
+  Scenario: I want to kill myself
+    Given a virologist and a full prot in a safe house
+    When the virologist steps on the safehouse
+    Then the virologist is in the safe house
